@@ -49,7 +49,6 @@
   "dependencies": {
     "cli-table": "^0.3.1",
     "commander": "^2.14.1",
-    "deasync": "^0.1.12",
     "inquirer": "^3.0.6",
     "js-yaml": "^3.9.1",
     "lodash": "^4.17.5",
@@ -57,9 +56,10 @@
     "merapi-proxy": "^0.1.3",
     "superagent": "^3.8.2",
     "uuid": "^3.2.1",
-    "zaun": "^2.0.0"
+    "zaun": "^2.0.1"
   },
   "devDependencies": {
+    "@types/inquirer": "0.0.43",
     "@types/js-yaml": "^3.9.0",
     "@types/lodash": "^4.14.104",
     "@types/node": "^7.0.55",
@@ -72,6 +72,7 @@
     "nyc": "^11.5.0",
     "sinon": "^3.2.0",
     "ts-node": "^3.3.0",
-    "typescript": "^2.7.2"
+    "tslint": "^5.11.0",
+    "typescript": "^3.1.6"
   }
 }
