@@ -1,10 +1,6 @@
 
-import { Component, JsonObject, IHash, Config, Json } from "merapi";
-import { v4 as uuid } from "uuid";
-import { ICompile, IHelper, ITester } from "interfaces/main";
-const colors = require("colors");
-const inquirer = require("inquirer");
-const path = require("path");
+import { IHelper } from "interfaces/main";
+import { Component, JsonObject } from "merapi";
 
 export default class Session extends Component {
     private defaultDeploymentId = "f223c9e0-6ba1-434d-8313-a9f18ca364bd";
